requirements.txt
alembic==1.8.1
click==8.1.2
colorama==0.4.6
Flask==2.2.2
Flask-Migrate==2.5.2
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
PyJWT==1.7.1
SQLAlchemy==1.4.44
Werkzeug==2.2.2
zipp==3.8.0